GPG Update

+++ Dockerfile
@@ -2,7 +2,7 @@ FROM tiredofit/alpine:3.4
 MAINTAINER Dave Conroy <dave at tiredofit dot ca>
 
 ### Set Nginx Version Number
-   ENV NGINX_VERSION 1.13.2
+   ENV NGINX_VERSION 1.13.4
 
 ### Install Nginx
    RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 && \
@@ -63,22 +63,31 @@ MAINTAINER Dave Conroy <dave at tiredofit dot ca>
 			pcre-dev \
 			zlib-dev \
 			linux-headers \
-                        tar \ 
+            tar \ 
 			gnupg \
 			libxslt-dev \
 			gd-dev \
 			geoip-dev \
 			perl-dev \
 			openldap-dev && \
-	mkdir -p /usr/src/nginx-auth-ldap /www /www/logs/nginx && \
-
+	    mkdir -p /usr/src/nginx-auth-ldap /www /www/logs/nginx && \
         curl https://codeload.github.com/kvspb/nginx-auth-ldap/tar.gz/master | tar xz --strip=1 -C /usr/src/nginx-auth-ldap && \
-	curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz -o nginx.tar.gz && \
+	    curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz -o nginx.tar.gz && \
         curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz.asc  -o nginx.tar.gz.asc && \
         export GNUPGHOME="$(mktemp -d)" && \
-        gpg --keyserver ha.pool.sks-keyservers.net --recv-keys "$GPG_KEYS" && \
-        gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz && \
-        rm -r "$GNUPGHOME" nginx.tar.gz.asc && \
+	    found=''; \
+			for server in \
+				ha.pool.sks-keyservers.net \
+				hkp://keyserver.ubuntu.com:80 \
+				hkp://p80.pool.sks-keyservers.net:80 \
+				pgp.mit.edu \
+			; do \
+				echo "Fetching GPG key $GPG_KEYS from $server"; \
+				gpg --keyserver "$server" --keyserver-options timeout=10 --recv-keys "$GPG_KEYS" && found=yes && break; \
+		    done; \
+	    test -z "$found" && echo >&2 "error: failed to fetch GPG key $GPG_KEYS" && exit 1; \
+	    gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz && \
+	    rm -r "$GNUPGHOME" nginx.tar.gz.asc && \
         mkdir -p /usr/src && \
         tar -zxC /usr/src -f nginx.tar.gz && \
         rm nginx.tar.gz && \
